Introduce scroll view controller with top separator
