+++ Sources/Charcoal/Filters/FilterViewController.swift
@@ -11,7 +11,7 @@ protocol FilterViewControllerDelegate: AnyObject {
     func filterViewControllerWillEndTextEditing(_ viewController: FilterViewController)
 }
 
-class FilterViewController: UIViewController, FilterBottomButtonViewDelegate {
+class FilterViewController: ScrollViewController, FilterBottomButtonViewDelegate {
     // MARK: - Public properties
 
     weak var delegate: FilterViewControllerDelegate?
@@ -21,17 +21,6 @@ class FilterViewController: UIViewController, FilterBottomButtonViewDelegate {
     // MARK: - Private properties
 
     lazy var bottomButtonBottomConstraint = bottomButton.bottomAnchor.constraint(equalTo: view.bottomAnchor)
-
-    private lazy var topSeparatorView: UIView = {
-        let view = UIView(withAutoLayout: true)
-        view.backgroundColor = .white
-        view.layer.masksToBounds = false
-        view.layer.shadowColor = UIColor.black.cgColor
-        view.layer.shadowOpacity = 0.7
-        view.layer.shadowOffset = CGSize(width: 0, height: 1)
-        view.layer.shadowRadius = 3
-        return view
-    }()
 
     private(set) lazy var bottomButton: FilterBottomButtonView = {
         let view = FilterBottomButtonView()
@@ -59,7 +48,6 @@ class FilterViewController: UIViewController, FilterBottomButtonViewDelegate {
         super.viewDidLoad()
         view.backgroundColor = .milk
         setup()
-        hideTopSeparator()
 
         let gestureRecognizer = UIScreenEdgePanGestureRecognizer()
         gestureRecognizer.delegate = self
@@ -109,48 +97,19 @@ class FilterViewController: UIViewController, FilterBottomButtonViewDelegate {
     // MARK: - Setup
 
     private func setup() {
-        view.addSubview(topSeparatorView)
         view.addSubview(bottomButton)
 
         NSLayoutConstraint.activate([
-            topSeparatorView.topAnchor.constraint(equalTo: view.topAnchor),
-            topSeparatorView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            topSeparatorView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
-            topSeparatorView.heightAnchor.constraint(equalToConstant: 1),
-
             bottomButton.leadingAnchor.constraint(equalTo: view.leadingAnchor),
             bottomButton.trailingAnchor.constraint(equalTo: view.trailingAnchor),
             bottomButton.topAnchor.constraint(lessThanOrEqualTo: view.bottomAnchor),
         ])
     }
 
-    // MARK: - Top separator
-
-    private func showTopSeparator() {
-        view.bringSubviewToFront(topSeparatorView)
-        topSeparatorView.isHidden = false
-    }
-
-    private func hideTopSeparator() {
-        topSeparatorView.isHidden = true
-    }
-
     // MARK: - FilterBottomButtonViewDelegate
 
     func filterBottomButtonView(_ filterBottomButtonView: FilterBottomButtonView, didTapButton button: UIButton) {
         delegate?.filterViewControllerDidPressBottomButton(self)
-    }
-}
-
-// MARK: - UIScrollViewDelegate
-
-extension FilterViewController: UIScrollViewDelegate {
-    func scrollViewDidScroll(_ scrollView: UIScrollView) {
-        if scrollView.contentOffset.y + scrollView.contentInset.top > 0 {
-            showTopSeparator()
-        } else {
-            hideTopSeparator()
-        }
     }
 }
